\name{torsaseuclid-package}
\alias{torsaseuclid-package}
\alias{torsaseuclid}
\docType{package}
\title{
\packageTitle{torsaseuclid}
}
\description{
\packageDescription{torsaseuclid}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{torsaseuclid}
\packageIndices{torsaseuclid}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{torsaseuclid}

Maintainer: \packageMaintainer{torsaseuclid}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
